Freenove 2.8 board support, MCP tools, backlight/audio/wake word improvements
- Add Freenove ESP32-S3 Display 2.8 LCD board
- Add MCP_TOOLS_GUIDE documentation
- Improve backlight control with AXP2101 support
- Fix audio service and codec initialization
- Fix AFE wake word and audio processor
- Fix WebSocket protocol
- Update sdkconfig for ESP32-S3
- Add AGENTS.md for OpenCode sessions

+++ main/mcp_server.h
@@ -332,11 +332,11 @@ private:
 
     void ParseCapabilities(const cJSON* capabilities);
 
-    void ReplyResult(int id, const std::string& result);
-    void ReplyError(int id, const std::string& message);
+    void ReplyResult(long long id, const std::string& result);
+    void ReplyError(long long id, const std::string& message);
 
-    void GetToolsList(int id, const std::string& cursor, bool list_user_only_tools);
-    void DoToolCall(int id, const std::string& tool_name, const cJSON* tool_arguments);
+    void GetToolsList(long long id, const std::string& cursor, bool list_user_only_tools);
+    void DoToolCall(long long id, const std::string& tool_name, const cJSON* tool_arguments);
 
     std::vector<McpTool*> tools_;
 };
